Extract region config into js/config.js
All Berlin-specific values (mapCenter, mapExtent, zoom levels, GeoServer
URL and workspace) are now consolidated in a single config file. To deploy
for a different city, only config.js needs to be edited.
