+++ js/config.js
@@ -1,0 +1,9 @@
+// Region configuration — edit this file to deploy for a different city/region
+
+export const mapCenter = [13.3888, 52.5170];                    // [lon, lat]
+export const mapExtent = [12.9494, 52.2577, 13.8126, 52.7928]; // [minLon, minLat, maxLon, maxLat]
+export const mapZoom = 12;
+export const mapMinZoom = 12;
+
+export const geoServer = 'https://osmbln.uber.space/';
+export const geoServerWorkspace = 'spielplatzkarte';
